add react plotly
