+++ package.json
@@ -22,10 +22,12 @@
     "node-sass": "^4.14.1",
     "npm-run-all": "^4.1.5",
     "ol": "^6.4.3",
+    "plotly.js-cartesian-dist": "^1.57.1",
     "postcss": "^7.0.32",
     "postcss-cli": "^7.1.1",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
+    "react-plotly.js": "^2.5.0",
     "react-scripts": "3.4.1",
     "redux-bundler": "^27.0.2",
     "redux-bundler-react": "^1.2.0",
