+++ src/app-common/plotly/A2Wplot.js
@@ -1,0 +1,6 @@
+import Plotly from "plotly.js-cartesian-dist";
+import createPlotlyComponent from "react-plotly.js/factory";
+
+const A2WPlot = createPlotlyComponent(Plotly);
+
+export default A2WPlot;
